requirements.txt
Django==3.1.10
PyYAML==5.4.1
Pygments==2.9.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
uritemplate==3.0.1
